🔧 Modify : replace class name
