+++ src/main/java/com/example/webtemplate/common/response/BaseResponse.java
@@ -1,10 +1,18 @@
 package com.example.webtemplate.common.response;
 
-import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
-public class BaseResponse {
+import com.example.webtemplate.common.ResponseMessage;
 
-    public record IndexResponse(List<Long> index) {
+public class BaseResponse<T> {
+    public record ResponseWrapper<T>(String message, T data) {
+    }
 
+    public static <T> ResponseEntity<ResponseWrapper<T>> of(int status, T data) {
+        HttpStatus httpStatus = HttpStatus.valueOf(status);
+        String message = ResponseMessage.of(status).getMessage();
+        var body = new ResponseWrapper<>(message, data);
+        return new ResponseEntity<>(body, httpStatus);
     }
 }
