+++ src/main/java/com/example/webtemplate/profile/UserProfileController.java
@@ -5,9 +5,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.webtemplate.common.response.BaseResponse.IndexResponse;
-import com.example.webtemplate.common.response.StandardResponse;
-import com.example.webtemplate.common.response.StandardResponse.ResponseWrapper;
+import com.example.webtemplate.common.response.StandardResponse.IndexResponse;
+import com.example.webtemplate.common.response.BaseResponse;
+import com.example.webtemplate.common.response.BaseResponse.ResponseWrapper;
 
 @RestController
 @RequestMapping("/profiles")
@@ -26,6 +26,6 @@ class UserProfileController {
   ResponseEntity<ResponseWrapper<IndexResponse>> accountIndex() {
     var accountIds = service.getAllUserProfileIds();
     var data = new IndexResponse(accountIds);
-    return StandardResponse.of(200, data);
+    return BaseResponse.of(200, data);
   }
 }
